Move gitlab modules to own subdir
