+++ lib/ansible/modules/source_control/gitlab/gitlab_deploy_key.py
