+++ test/sanity/ignore.txt
@@ -208,8 +208,8 @@ lib/ansible/module_utils/gcp.py future-import-boilerplate
 lib/ansible/module_utils/gcp.py metaclass-boilerplate
 lib/ansible/module_utils/gcp_utils.py future-import-boilerplate
 lib/ansible/module_utils/gcp_utils.py metaclass-boilerplate
-lib/ansible/module_utils/gitlab.py future-import-boilerplate
-lib/ansible/module_utils/gitlab.py metaclass-boilerplate
+lib/ansible/module_utils/gitlab/gitlab.py future-import-boilerplate
+lib/ansible/module_utils/gitlab/gitlab.py metaclass-boilerplate
 lib/ansible/module_utils/hwc_utils.py future-import-boilerplate
 lib/ansible/module_utils/hwc_utils.py metaclass-boilerplate
 lib/ansible/module_utils/infinibox.py future-import-boilerplate
@@ -5717,8 +5717,8 @@ test/units/modules/remote_management/oneview/test_oneview_san_manager.py future-
 test/units/modules/remote_management/oneview/test_oneview_san_manager.py metaclass-boilerplate
 test/units/modules/remote_management/oneview/test_oneview_san_manager_info.py future-import-boilerplate
 test/units/modules/remote_management/oneview/test_oneview_san_manager_info.py metaclass-boilerplate
-test/units/modules/source_control/gitlab.py future-import-boilerplate
-test/units/modules/source_control/gitlab.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/gitlab.py future-import-boilerplate
+test/units/modules/source_control/gitlab/gitlab.py metaclass-boilerplate
 test/units/modules/source_control/test_bitbucket_access_key.py future-import-boilerplate
 test/units/modules/source_control/test_bitbucket_access_key.py metaclass-boilerplate
 test/units/modules/source_control/test_bitbucket_pipeline_key_pair.py future-import-boilerplate
@@ -5727,18 +5727,18 @@ test/units/modules/source_control/test_bitbucket_pipeline_known_host.py future-i
 test/units/modules/source_control/test_bitbucket_pipeline_known_host.py metaclass-boilerplate
 test/units/modules/source_control/test_bitbucket_pipeline_variable.py future-import-boilerplate
 test/units/modules/source_control/test_bitbucket_pipeline_variable.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_deploy_key.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_deploy_key.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_group.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_group.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_hook.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_hook.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_project.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_project.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_runner.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_runner.py metaclass-boilerplate
-test/units/modules/source_control/test_gitlab_user.py future-import-boilerplate
-test/units/modules/source_control/test_gitlab_user.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_deploy_key.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_deploy_key.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_group.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_group.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_hook.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_hook.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_project.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_project.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_runner.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_runner.py metaclass-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_user.py future-import-boilerplate
+test/units/modules/source_control/gitlab/test_gitlab_user.py metaclass-boilerplate
 test/units/modules/storage/hpe3par/test_ss_3par_cpg.py future-import-boilerplate
 test/units/modules/storage/hpe3par/test_ss_3par_cpg.py metaclass-boilerplate
 test/units/modules/storage/netapp/test_na_elementsw_cluster_config.py future-import-boilerplate
